ai-startup-evaluator/.backup_ai-startup-evaluator_20250921_113511/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
google-adk==0.2.0
google-cloud-storage==2.10.0
google-cloud-sql-connector==1.4.3
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
python-pptx==0.6.23
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
SpeechRecognition==3.10.0
pydub==0.25.1
